Fixed server gradual server crash
The previous implementation had an issue where the array of clients
weren't being used effectively. If a connection re-used a client
handler, the server crashed.

This was fixed by using a `ThreadPoolExecutor` to manage the threads. The
thread executor allows setting a max limit, which throws an exception if
exceeded. `ThreadPoolExecutor` also allows queuing up connections, which
means we could have more than 4 connections in the future.